+++ src/xenia/cpu/backend/backend.h
@@ -46,7 +46,7 @@ using SimpleGuestTrampolineProc = void (*)(ppc::PPCContext*);
 // much host code that produced, and which guest instruction of the function it
 // belongs to, so the coverage counters can weight it.
 struct SequenceSample {
-  uint32_t key;
+  uint64_t key;
   uint32_t guest_index;
   uint32_t host_bytes;
 };
@@ -156,7 +156,7 @@ class Backend {
 
   // Renders a sequence selection key as "OPCODE_NAME dest src1 src2 src3" for
   // the profiler dump. The key layout is private to each backend.
-  virtual std::string FormatSequenceKey(uint32_t key) const { return ""; }
+  virtual std::string FormatSequenceKey(uint64_t key) const { return ""; }
 
  protected:
   Processor* processor_ = nullptr;
